requirements.txt
cffi==1.17.1
cryptography==43.0.1
greenlet==3.1.1
pycparser==2.22
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
SQLAlchemy==2.0.35
typing_extensions==4.12.2
Pillow==11.1.0
